Split podUtil.spec.ts in the same way as podUtil.ts was split
